requirements.txt
redis
requests
fastapi
python-dotenv
